KarmaRanking/requirements.txt
astroid==2.2.5
autopep8==1.4.4
cachetools==4.0.0
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
google-api-python-client==1.7.11
google-auth==1.11.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
httplib2==0.17.0
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
oauthlib==3.1.0
praw==6.3.1
prawcore==1.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pylint==2.3.1
python-dotenv==0.10.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.12.0
typed-ast==1.4.0
update-checker==0.16
uritemplate==3.0.1
urllib3==1.25.3
websocket-client==0.56.0
wrapt==1.11.2
